refactor: change plan-selection to plan-selection-deprecated
