+++ src/i18n/en.js
@@ -29,7 +29,7 @@ const urlSiteFromLogin = `${urlSite}/en/`;
 const urlSiteFromForgot = `${urlSite}/en/`;
 const urlSiteContact = `${urlSite}/en/contact/`;
 const urlControlPanelMain = `${urlControlPanel}/ControlPanel`;
-const urlPlanSelection = `/plan-selection`;
+const urlPlanSelection = `/plan-selection-deprecated`;
 const urlMasterSubscriber = `${urlDopplerLegacy}/Lists/MasterSubscriber/`;
 const urlCreateSubscriberList = `${urlDopplerLegacy}/Lists/List`;
 const urlDraft = `${urlDopplerLegacy}`;
